We're done with the login function. Kind of.
I logged in, and it does in fact show my name, but It looks like the grades are actually rendered on the page in an iframe. So the grades are a separate page. So, I've tried POSTing and GETing to that page, with out the form data and headers, but I have been unsuccessful. I may have to take a different approach to scraping the data.