AYYYY WE FIXED THE LOGIN FUNCTION!!!
SO, what I'm doing now is POSTing to the Week view page, which is shown as the first page after you login. I do this because the data on this page is not in an iframe, and therefore I can scrape without any issues. Eventually, I have to figure out why I can't get onto /content pages, but I believe it has something to do with cookies. Next step: scraping. My future tastes like beautiful soup.