I got almost all the scraping code done!
We just have one interesting kink with the teacher list. It's adding extra spaces. Minus that, we're very close to completion. Also, I realized that when I got lazy earlier and disabled the getUserCreds function and directly set my username and password in the code, it has also been committed and saved into the code for the last couple commits! I'll eventually have to fix that, but fortunately this is a private repository. I know it is possible for me to edit old commits, though.