WE BASICALLY FINISHED WOOO
We have one issue: The lunches throw stuff off. You have a screenshot in your email showing the issue. Essentially, since there is no element for teacher name for lunches, lxml doesn't scrape a teacher, so when I'm printing the results, it prints the teacher of the next class. I have a comment with a potential solution.